+++ examples/react-server/src/features/bootstrap/plugin.ts
@@ -2,8 +2,8 @@ import fs from "node:fs";
 import { tinyassert, typedBoolean } from "@hiogawa/utils";
 import type { Manifest, PluginOption, ViteDevServer } from "vite";
 import { $__global } from "../../global";
+import { VIRTUAL_COPY_SERVER_CSS } from "../style/plugin";
 import { createVirtualPlugin } from "../utils/plugin";
-import { VIRTUAL_COPY_SERVER_CSS } from "./css";
 
 export const ENTRY_CLIENT_BOOTSTRAP = "virtual:entry-client-bootstrap";
